refactor map into several sub-modules
